+++ lib/ajax-datatables-rails/configuration.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module AjaxDatatablesRails
-  class Configuration
-    include ActiveSupport::Configurable
-
-    config_accessor(:nulls_last) { false }
-  end
-end